// mehr-new-structure/src/lib/api/models/permission.ts
/**
 * Generated by orval v7.14.0 🍺
 * Do not edit manually.
 * clinic reservation system api
 * OpenAPI spec version: 2.0.0
 */
import type { PermissionCategory } from './permissionCategory';

export interface Permission {
  name: string;
  title: string;
  translatedTitle: string;
  guardName: string;
  categoryName: string;
  category: PermissionCategory;
  categoryId: string;
  id: string;
  createdAt: string;
  updatedAt: string;
  deletedAt: string;
  creatorId: string;
}