// mehr-new-structure/src/lib/api/models/role.ts
/**
 * Generated by orval v7.14.0 🍺
 * Do not edit manually.
 * clinic reservation system api
 * OpenAPI spec version: 2.0.0
 */
import type { Permission } from './permission';

export interface Role {
  name: string;
  title: string;
  guardName: string;
  expireDate: string;
  permissions: Permission[];
  id: string;
  createdAt: string;
  updatedAt: string;
  deletedAt: string;
  creatorId: string;
}